fix slice_from_axis to wrap zarr in dask to keep lazy behavior (#8260)
# References and relevant issues
Closes: https://github.com/napari/napari/issues/8008
see also discussion at:
https://napari.zulipchat.com/#narrow/channel/212875-general/topic/handling.20zarr.2Fdask.2Fetc.20when.20splitting.20or.20merging.20layers/with/532610558

# Description
So the issue is that when you slice a zarr array, it returns a numpy
array.
This means that the slice_from_axis function breaks the laziness for
zarr arrays and loads everything into memory.
This is a known issue with zarr:
https://github.com/zarr-developers/zarr-python/discussions/1603

In this PR, I add code to check if an array is a zarr array and if so,
wrap it in a dask.array, so that lazy behavior is preserved. I got some
help from Claude to come up with a way to check for being a zarr without
importing -- if someone is passing a zarr, they have zarr obviously, but
napari does not depend on zarr so using the simple `isinstance(array,
zarr.Array)` would result in errors.
I'm open to other solutions, could try/except obviously, but this seemed
better?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>